internal/binary: new package
